requirements.txt
flask>=2.3.3
werkzeug>=2.3.7
requests>=2.31.0
line-bot-sdk>=3.1.0
openai>=1.3.0
httpx>=0.25.0
pytest>=7.3.1
pytest-cov>=4.1.0
flask-talisman>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0